Add files via upload
This is a code to print the HMS time of the ONOFF mode with 5 min ON the target, 1 min shift from ON to OFF, 5 min OFF the target, 1 min shift back as one ONOFF cycle. And add the 10 seconds noise at the beginning of each ON and OFF. Then the noise results will be more close to the flux during the observation. It is also said that during the shifting of the telescope between ON and OFF, the noise or the whole system is not steady, so we should not make use of these data. Here I add 20s noise during the shift to test if the noise calibration values are steady.